2_ai_image_captioner/requirements.txt
asttokens==2.4.1
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
bs4==0.0.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.7
decorator==5.1.1
executing==2.1.0
filelock==3.13.1
Flask==3.0.3
fonttools==4.54.1
fsspec==2024.2.0
h11==0.14.0
huggingface-hub==0.26.2
idna==3.10
ImageHash==4.3.1
imageio==2.36.0
ipykernel==6.29.5
ipython==8.29.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
lazy_loader==0.4
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.3
outcome==1.3.0.post0
packaging==24.1
parso==0.8.4
pillow==10.2.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.1.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.18.0
pyparsing==3.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyWavelets==1.7.0
# pywin32==308
PyYAML==6.0.2
pyzmq==26.2.0
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
scikit-image==0.24.0
scipy==1.14.1
selenium==4.26.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
stack-data==0.6.3
sympy==1.13.1
tifffile==2024.9.20
tokenizers==0.20.1
# torch==2.5.1+cu124
# torchaudio==2.5.1+cu124
# torchvision==0.20.1+cu124
tornado==6.4.1
tqdm==4.66.6
traitlets==5.14.3
transformers==4.46.1
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.9.0
urllib3==2.2.3
wcwidth==0.2.13
websocket-client==1.8.0
Werkzeug==3.1.0
wsproto==1.2.0
